+++ src/move-to-front-transform.hpp
@@ -2,12 +2,12 @@
  * @file move-to-front-transform.hpp
  * @author Juho Röyskö
  * @brief Header file for move-to-front-transform.cpp
- * @version 0.1
- * @date 2021-09-17
+ * @version 0.2
+ * @date 2021-09-20
  */
 #pragma once
 #include <string>
 #include <vector>
 
-std::vector<int> MTFTEncode(std::string input);
-std::string MTFTDecode(std::vector<int> input);
+std::vector<uint8_t> MTFTEncode(std::string input);
+std::string MTFTDecode(std::vector<uint8_t> input);
